common: bwl: move public headers to include/
Similar like is done in other modules, put the public headers in the
include/ directory, with the same directory structure as where it is
installed.

This simplifies the installation command to just use DIRECTORY instead
of listing the files.

base_wlan_hal.cpp doesn't move, but it's #include is updated to use <>
instead of "" since the included file is no longer in the same
directory. Keep the <bwl/..> prefix to diffrenciate between include
files that can potentially have the same name.

While we're at it, also make the include directory a public include
directory, for the sake of internal consumers (of which there are none
at the moment).
