+++ agent/src/beerocks/monitor/mon_wlan_hal_loader.cpp
@@ -7,7 +7,7 @@
  */
 
 // Monitor HAL
-#include <beerocks/bwl/mon_wlan_hal.h>
+#include <bwl/mon_wlan_hal.h>
 
 #include <dlfcn.h> // Dynamic loading
