+++ agent/src/beerocks/slave/ap_wlan_hal_loader.cpp
@@ -7,7 +7,7 @@
  */
 
 // AP HAL
-#include <beerocks/bwl/ap_wlan_hal.h>
+#include <bwl/ap_wlan_hal.h>
 
 #include <dlfcn.h> // Dynamic loading
