+++ agent/src/beerocks/slave/backhaul_manager/sta_wlan_hal_loader.cpp
@@ -7,7 +7,7 @@
  */
 
 // Monitor HAL
-#include <beerocks/bwl/sta_wlan_hal.h>
+#include <bwl/sta_wlan_hal.h>
 
 #include <dlfcn.h> // Dynamic loading
