+++ common/beerocks/bwl/dummy/ap_wlan_hal_dummy.h
@@ -10,7 +10,7 @@
 #define _BWL_AP_WLAN_HAL_DUMMY_H_
 
 #include "base_wlan_hal_dummy.h"
-#include <ap_wlan_hal.h>
+#include <bwl/ap_wlan_hal.h>
 
 namespace bwl {
 namespace dummy {
