+++ common/beerocks/bwl/dummy/mon_wlan_hal_dummy.h
@@ -10,7 +10,7 @@
 #define _BWL_MON_WLAN_HAL_DUMMY_H_
 
 #include "base_wlan_hal_dummy.h"
-#include <mon_wlan_hal.h>
+#include <bwl/mon_wlan_hal.h>
 
 namespace bwl {
 namespace dummy {
