+++ common/beerocks/bwl/dummy/sta_wlan_hal_dummy.h
@@ -10,7 +10,7 @@
 #define _BWL_STA_WLAN_HAL_DUMMY_H_
 
 #include "base_wlan_hal_dummy.h"
-#include <sta_wlan_hal.h>
+#include <bwl/sta_wlan_hal.h>
 
 namespace bwl {
 namespace dummy {
