+++ common/beerocks/bwl/dwpal/ap_wlan_hal_dwpal.h
@@ -10,7 +10,7 @@
 #define _BWL_AP_WLAN_HAL_DWPAL_H_
 
 #include "base_wlan_hal_dwpal.h"
-#include <ap_wlan_hal.h>
+#include <bwl/ap_wlan_hal.h>
 
 namespace bwl {
 namespace dwpal {
