+++ common/beerocks/bwl/dwpal/mon_wlan_hal_dwpal.h
@@ -10,7 +10,7 @@
 #define _BWL_MON_WLAN_HAL_DWPAL_H_
 
 #include "base_wlan_hal_dwpal.h"
-#include <mon_wlan_hal.h>
+#include <bwl/mon_wlan_hal.h>
 
 namespace bwl {
 namespace dwpal {
