+++ common/beerocks/bwl/dwpal/sta_wlan_hal_dwpal.h
@@ -10,7 +10,7 @@
 #define _BWL_STA_WLAN_HAL_DWPAL_H_
 
 #include "base_wlan_hal_dwpal.h"
-#include <sta_wlan_hal.h>
+#include <bwl/sta_wlan_hal.h>
 
 namespace bwl {
 namespace dwpal {
