+++ common/beerocks/bwl/nl80211/ap_wlan_hal_nl80211.h
@@ -10,7 +10,7 @@
 #define _BWL_AP_WLAN_HAL_NL80211_H_
 
 #include "base_wlan_hal_nl80211.h"
-#include <ap_wlan_hal.h>
+#include <bwl/ap_wlan_hal.h>
 
 namespace bwl {
 namespace nl80211 {
