+++ common/beerocks/bwl/wav/ap_wlan_hal_wav.h
@@ -10,7 +10,7 @@
 #define _BWL_AP_WLAN_HAL_WAV_H_
 
 #include "base_wlan_hal_wav.h"
-#include <ap_wlan_hal.h>
+#include <bwl/ap_wlan_hal.h>
 
 namespace bwl {
 namespace wav {
