+++ common/beerocks/bwl/wav/mon_wlan_hal_wav.h
@@ -10,7 +10,7 @@
 #define _BWL_MON_WLAN_HAL_WAV_H_
 
 #include "base_wlan_hal_wav.h"
-#include <mon_wlan_hal.h>
+#include <bwl/mon_wlan_hal.h>
 
 namespace bwl {
 namespace wav {
