+++ common/beerocks/bwl/wav/sta_wlan_hal_wav.h
@@ -10,7 +10,7 @@
 #define _BWL_STA_WLAN_HAL_WAV_H_
 
 #include "base_wlan_hal_wav.h"
-#include <sta_wlan_hal.h>
+#include <bwl/sta_wlan_hal.h>
 
 namespace bwl {
 namespace wav {
