+++ common/beerocks/bwl/wav_fapi/fapi_attach_fsm.cpp
@@ -8,7 +8,7 @@
 
 #include "fapi_attach_fsm.h"
 
-#include <base_wlan_hal_types.h>
+#include <bwl/base_wlan_hal_types.h>
 
 #include <beerocks/bcl/beerocks_utils.h>
 #include <easylogging++.h>
